move info from request in SparqlRequest
